requirements.txt
pytorch
torchvision
numpy
matplotlib